pyproject.toml
[project]
name = "autogen"
version = "0.1.0"
description = "AutoGen 0.4 example project with custom agents and workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "autogen-agentchat>=0.4.5",
    "autogen-ext[openai]>=0.4.5",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.393",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.5",
    "pytest-env>=1.1.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
env = [
    "PYTHONPATH=src",
]

[tool.pyright]
include = ["src", "tests"]
typeCheckingMode = "basic"
useLibraryCodeForTypes = true
reportMissingImports = true
reportUnknownMemberType = false
